:art: - Rename package of the API error handlers and add mapping for httpMediaTypeNotAcceptable exception
